@@ -4,8 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "webpack": "node_modules/webpack-cli/bin/cli.js --config webpack.production.js",
-    "webpack-development": "node_modules/webpack-cli/bin/cli.js --config webpack.development.js",
+    "webpack": "webpack --config webpack.production.ts",
+    "webpack-development": "webpack --config webpack.development.ts",
     "relay-compiler": "babel-node -x .ts -- bin/generateSchemaAndRunRelayCompiler.ts"
   },
   "relay": {
@@ -30,22 +30,23 @@
     "@stylexjs/stylex": "^0.10.0",
     "graphiql": "^3.8.3",
     "graphql-tools": "^9.0.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-relay": "^16.2.0",
     "react-router-dom": "^6.22.0",
     "relay-runtime": "^16.2.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.0",
-    "@babel/node": "^7.23.9",
+    "@babel/core": "^7.27.1",
+    "@babel/node": "^7.27.1",
     "@babel/preset-env": "^7.24.0",
-    "@babel/preset-react": "^7.23.3",
-    "@babel/preset-typescript": "^7.23.3",
+    "@babel/preset-react": "^7.27.1",
+    "@babel/preset-typescript": "^7.27.1",
+    "@caseyjaymcguire/stylex-webpack-plugin": "^0.0.4",
     "@stylexjs/webpack-plugin": "^0.10.0",
     "@types/js-beautify": "^1.14.3",
-    "@types/react": "^18.2.0",
-    "@types/react-dom": "^18.2.0",
+    "@types/react": "^18.3.1",
+    "@types/react-dom": "^18.3.1",
     "@types/react-relay": "^14.1.3",
     "@types/relay-runtime": "^14.1.23",
     "babel-loader": "^8.0.6",
@@ -59,8 +60,8 @@
     "style-loader": "^4.0.0",
     "ts-loader": "^9.5.0",
     "typescript": "^5.3.3",
-    "webpack": "^5.90.1",
+    "webpack": "^5.97.1",
     "webpack-cli": "^5.1.4",
-    "webpack-merge": "^4.2.2"
+    "webpack-merge": "^6.0.1"
   }
 }
