@@ -1,8 +1,11 @@
-const merge = require("webpack-merge");
-const common = require("./webpack.common.js");
-const path = require('path');
+import common from "./webpack.common";
 
-module.exports = merge(common, {
+import merge from "webpack-merge";
+
+
+import path from "path";
+
+export default merge(common, {
   watchOptions: {
     poll: true
   },
@@ -14,7 +17,8 @@ module.exports = merge(common, {
     filename: '[name].bundle.js',
     // I couldn't figure out how to get Spring Boot or Intellij to deploy changed webpack bundles instantaneously
     // so we just have webpack deploy directly into the build folder here.
-    path: path.resolve(__dirname, 'build/resources/main/static/bundles')
+    path: path.resolve(__dirname, 'build/resources/main/static/bundles'),
+    module: true
   },
   stats: {
     errorDetails: true
